Enrichment Scores

This function computes an enrichment score for each column vector in the input matrix x. See the Details section below for further information.

Usage

colEnrichment(
  x,
  rows = NULL,
  na.rm = TRUE,
  method = c("ssgsea", "gsva", "plage", "zscore"),
  verbose = FALSE,
  ...
)

Arguments

x

a numerical matrix features-by-samples

rows

(optional) numerical vector giving the rows in x or character vector matching the row names in x to operate over. If missing or rows = NULL, all the rows in x are considered for the computation of the measures

na.rm

logical, whether to remove NA values from x before computation

method

character string, the name of the method to be used for computing the enrichment scores. Current available options are:

'ssgsea'

Single Sample Gene Set Enrichment Analysis

'gsva'

Gene Set Variation Analysis

'plage'

Pathway Level Analysis of Gene Expression

'zscore'

Z-Score

verbose

Gives information about each calculation step. Default: TRUE.

...

further arguments to the method-specific parameters object used in gsva

Value

A numerical vector containing the computed score for each column in x.

Details

It is a wrapper to gsva function.

See also

Author

Alessandro Barberis