Compute Measure

This function computes statistical measure(s) from an input vector or matrix. It is an internal function, and it is not intended to be called by users.

Usage

computeMeasure(
  x,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqm", "iqr", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  na.rm = TRUE,
  ...
)

Arguments

x

(named) numerical vector or a matrix features-by-samples

score

character string indicating the statistical measure to compute

na.rm

logical, whether to remove NA values from x before computation

...

further arguments to score

Value

A numerical vector representing the computed measure(s).

Details

If x is a vector, the output is a single value, if x is a matrix, the output is a vector of length ncol(x).

A default NA value is returned if the score can't be computed, or if i values are not present in x. If x is a matrix, the output is a vector of NA values.

See also

Author

Alessandro Barberis