Generic Scorer Function

This function computes a summary score from an input vector or from each column vector of an input matrix.

Usage

computeScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = T,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = FALSE,
  ...
)

sumScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

weightedSumScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

meanScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

trimmedMeanScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

weightedMeanScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

medianScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

modeScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

midrangeScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

midhingeScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

trimeanScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

iqrScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

iqmScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

madScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

aadScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

ssgseaScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

gsvaScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

plageScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

zscoreScore(
  x,
  i = NULL,
  w = NULL,
  na.rm = TRUE,
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = F,
  ...
)

Arguments

x

(named) numerical vector or matrix

i

(optional) numerical vector giving the (row) position in x or character vector matching the (row) names in x. If missing or i = NULL, the entire x is considered for the computation of the score

w

numerical vector of weights the same length as i. It is used for the computation of the weighted scores

na.rm

logical, whether to remove NA values from x before computation

score

character string indicating the summary score to compute

transform.fun

function to transform the data. If provided, x is transformed (x = transform.fun(x, transform.args)) before the computation of the scores. See getDataTransformer for further details about built-in options

transform.args

list of parameters to the data transformation function

transform.sub

logical, whether to transform x after it is subset for i (used to speedup computation). Default is FALSE, meaning the transformation would be applied directly to x provided in input

...

further arguments to score

Value

A numerical value representing the computed score. A default NA value is returned if the score can't be computed, or if i values are not present in x.

Details

The function handles vector or matrix input by calling computeVectorMeasure or computeColMeasures, respectively.

Internally, computeColMeasures uses these functions to compute the measures:

"sum"

colSummations

"weightedSum"

colWeightedSums

"mean"

colArithmeticMeans

"trimmedMean"

colTrimmedMeans

"weightedMean"

colWeightedArithmeticMeans

"median"

colMidpoints

"mode"

colModes

"midrange"

colMidranges

"midhinge"

colMidhinges

"trimean"

colTrimeans

"iqr"

colIQRs

"iqm"

colIQMs

"mad"

colMADs

"aad"

colAADs

"ssgsea"

colSsgsea

"gsva"

colGsva

"plage"

colPlage

"zscore"

colZscore

computeVectorMeasure uses instead these functions:

"sum"

summation

"weightedSum"

weightedSum

"mean"

arithmeticMean

"trimmedMean"

trimmedMean

"weightedMean"

weightedMean

"median"

midpoint

"mode"

modalValue

"midrange"

midrange

"midhinge"

midhinge

"trimean"

trimean

"iqr"

iqr

"iqm"

IQM

"mad"

MAD

"aad"

AAD

"ssgsea"

ssgsea

"gsva"

gsva

"plage"

plage

"zscore"

zscore

Functions

sumScore(): returns the sum score(s).

weightedSumScore(): returns the weighted sum score(s).

meanScore(): returns the arithmentic mean score(s).

trimmedMeanScore(): returns the trimmed mean score(s).

weightedMeanScore(): returns the weighted mean score(s).

medianScore(): returns the median score(s).

modeScore(): returns the mode score(s).

midrangeScore(): returns the midrange score(s).

midhingeScore(): returns the midhinge score(s).

trimeanScore(): returns the trimean score(s).

iqrScore(): returns the interquartile range score(s).

iqmScore(): returns the interquartile mean score(s).

madScore(): returns the median absolute deviation score(s).

aadScore(): returns the average absolute deviation score(s).

ssgseaScore(): returns the single sample gene set enrichment analysis (ssGSEA) score(s).

gsvaScore(): returns the gene set variation analysis (GSVA) score(s).

plageScore(): returns the pathway level analysis of gene expression (PLAGE) score(s).

zscoreScore(): returns the z-score(s).

See also

Author

Alessandro Barberis