Logger Class Constructor

Constructor for the object.

Usage

createLogger(
  path = character(),
  con = NULL,
  verbose = TRUE,
  level = c("INFO", "DEBUG", "TRACE", "ALL", "OFF")
)

Arguments

path

a path to the log file to use

con

(optional) connection to a file

verbose

logical, whether R should report extra information on progress

level

the log level of the logger. Five levels are supported:

"OFF"

indicates the logger is inactive

"INFO"

indicates informational messages

"DEBUG"

indicates fine-grained informational messages

"TRACE"

indicates finer-grained informational events than the "DEBUG"

"ALL"

all levels are considered

Levels are considered ordered: OFF < INFO < DEBUG < TRACE < ALL

Value

A Logger object.

Details

A log request in a Logger of level l is enabled if the level of the request is <= l.

See also

Author

Alessandro Barberis