Looper

Common interface to run a for loop sequentially or in parallel. The core of the loop is defined by the function fun.

Usage

forLoop(
  n.iter,
  cores = 1L,
  .inorder = TRUE,
  fun,
  ...,
  logger = NULL,
  verbose = TRUE
)

Arguments

n.iter

number of iterations of the loop

cores

number of cores to use for parallel execution.

.inorder

logical flag indicating whether the .combine function requires the task results to be combined in the same order that they were submitted. If the order is not important, then it setting .inorder to FALSE can give improved performance. The default value is `TRUE.

fun

A function to be executed inside the loop. It should contain an iloop parameter.

...

further arguments to fun

logger

a Logger

verbose

logical, indicating whether R should report extra information on progress

Author

Alessandro Barberis