Step Function

This function implements a Step Function. See the Details section below for further information.

Usage

stepFunctionForVector(x, y = c(-1, 0, 1), thr = 0)

Arguments

x

(named) numerical vector

y

numerical vector of length 3, the values to use for the piecewise function. The first value is used for the left interval, the second value is used when x == threshold, and the third value is used for the right interval

thr

numerical value, the threshold to use to divide values in x in two intervals

Value

A numerical vector, containing the output of the step function for each element of x. The default settings of thr and y makes this function equivalent to a sign function.

Details

This function implements a piecewise function that is defined as follows:

$$SF = f(x) = \left\{ \begin{array}{l} y_{1}, \quad \textrm{if} \quad x > threshold \\ y_{2}, \quad \textrm{if} \quad x = threshold \\ y_{3}, \quad \textrm{if} \quad x < threshold \end{array} \right\} $$

Author

Alessandro Barberis