Compute the Summary Scores

This function computes summary score(s) of the signature i in input considering each column vector in the input matrix x.

A parallel execution to speed up the computation on a multi-core machine can be run by setting the argument cores with a number greater than 1.

Usage

computeSigScores(
  x,
  i = NULL,
  na.rm = TRUE,
  scores = c("mean", "trimmedMean", "weightedMean", "median", "mode", "midrange",
    "midhinge", "trimean", "briskow", "reviewedBriskow", "IQM", "weightedSum", "ssGSEA",
    "gsva", "plage", "zscore"),
  sampling = c("none", "permutation", "bootstrap"),
  n.repeat = 10L,
  cores = 1L,
  args = NULL
)

Arguments

x

numerical matrix, features-by-samples

i

(optional) numerical vector giving the rows in x or character vector matching the row names in x If missing or i = NULL, all the rows in x are considered for the computation of the score

na.rm

logical, whether to remove NA values before computation

scores

character vector, indicating the summary score(s) to compute

sampling

character string, indicating whether to compute the scores using the provided data (sampling = "none", default) or whether to sample the data.

Three options are available:

none

use x as it is

permutation

random sampling without replacement

bootstrap

random sampling with replacement

n.repeat

integer, number of repeated samples to generate

cores

number of cores to use for parallel execution.

args

named list, where the names must match the scores. Each element in the list is another list containing the arguments to pass to the function used for computing the named score. For example, args = list(trimmedMean = list(trim = 0.4)) indicates to use trim = 0.4 when computing the trimmed mean score

Value

A data frame containing the computed score(s) for each sample. Each row corresponds to a different sample.

If sampling = "random" or sampling = "bootstrap", the data frame contains a column with the run information.

The two columns containing the run/sample information are:

sampleID

the name of the sample

run

integer indicating in which run - out of the n.repeat - was computed the score

See also

Author

Alessandro Barberis

Examples

if (FALSE) {
#set seed for reproducibility
set.seed(seed = 5381L)

#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Compute all scores
computeSigScores(
 x = x,
 i = rownames(x)[1:10]
)

#Compute one score
computeSigScores(
 x = x,
 i = rownames(x)[1:10],
 scores = 'mean'
)

#Compute one score passing an argument
computeSigScores(
 x = x,
 i = rownames(x)[1:10],
 scores = 'trimmedMean',
 args = list(trimmedMean = list(trim = 0.2))
)

#Compute scores with permutation
computeSigScores(
 x        = x,
 i        = rownames(x)[1:10],
 sampling = "permutation",
 n.repeat = 10
)

#Compute scores with bootstrap
computeSigScores(
 x        = x,
 i        = rownames(x)[1:10],
 sampling = "bootstrap",
 n.repeat = 10
)
}