Get Scorer

This function is a dispatcher for the score function selected in input.

Usage

getScorer(
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore")
)

Arguments

score

character string, one of the supported measures

Value

A scoring function:

"sum"

sumScorer

"weightedSum"

weightedSumScorer

"mean"

meanScorer

"trimmedMean"

trimmedMeanScorer

"weightedMean"

weightedMeanScorer

"median"

medianScorer

"mode"

modeScorer

"midrange"

midrangeScorer

"midhinge"

midhingeScorer

"trimean"

trimeanScorer

"iqr"

iqrScorer

"iqm"

iqmScorer

"mad"

madScorer

"aad"

aadScorer

"ssgsea"

ssgseaScorer

"gsva"

gsvaScorer

"plage"

plageScorer

"zscore"

zscoreScorer

Details

The scoring functions are created via the generic scorer genericScorer which handles vector or matrix input by calling computeColMeasures.

Internally, computeColMeasures uses these functions to compute the measures: colSummations, colWeightedSums, colArithmeticMeans, colTrimmedMeans, colWeightedArithmeticMeans, colMidpoints, colModes, colMidranges, colMidhinges, colTrimeans, colIQRs, colIQMs, colMADs, colAADs, colSsgsea, colGsva, colPlage, colZscore

Author

Alessandro Barberis