Heatmap

This function generates a ComplexHeatmap object.

It is a wrapper around the Heatmap function.

Usage

heatMap(
  data,
  na_col = "grey",
  color_space = "LAB",
  rect_gp = grid::gpar(col = "white"),
  col = NULL,
  column_title = character(0),
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  ...
)

Arguments

data

data frame, output of computeSigScores

na_col

Color for NA values.

color_space

The color space in which colors are interpolated. Only used if matrix is numeric and col is a vector of colors. Pass to colorRamp2.

rect_gp

Graphic parameters for drawing rectangles (for heatmap body). The value should be specified by gpar and fill parameter is ignored.

col

A vector of colors if the color mapping is discrete or a color mapping function if the matrix is continuous numbers (should be generated by colorRamp2). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to ColorMapping. For more details and examples, please refer to https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors .

column_title

Title on the column.

cluster_rows

If the value is a logical, it controls whether to make cluster on rows. The value can also be a hclust or a dendrogram which already contains clustering. Check https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#clustering .

cluster_columns

Whether make cluster on columns? Same settings as cluster_rows.

...

further arguments to internal function call

Value

A ComplexHeatmap object.

Author

Alessandro Barberis